requirements.txt
python-dotenv==1.1.0
python-telegram-bot==13.15
requests==2.32.3
urllib3==1.26.16